Moved ping utility to correct folder
