+++ BaseStation/PingMonitor/PingMonitor/MainWindow.xaml
@@ -1,0 +1,11 @@
+﻿<Window x:Class="HuskyRobotics.UI.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:HuskyRobotics.UI"
+        mc:Ignorable="d"
+        Title="MainWindow" Height="350" Width="525">
+    <!-- I'm using DockPanels here simply because I like them, other layouts work fine -->
+    <local:PingMonitor/>
+</Window>
