+++ BaseStation/PingMonitor/PingMonitor/MainWindow.xaml.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace HuskyRobotics.UI
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// 
+    /// This class represents the VeiwModel
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+
+        public MainWindow()
+        {
+            InitializeComponent();
+
+        }
+
+
+    }
+}
