--- UI/PingMonitor/PingMonitor.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2010
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PingMonitor", "PingMonitor\PingMonitor.csproj", "{BB653C08-22A9-4440-82F0-804067B83DFD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BB653C08-22A9-4440-82F0-804067B83DFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BB653C08-22A9-4440-82F0-804067B83DFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BB653C08-22A9-4440-82F0-804067B83DFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BB653C08-22A9-4440-82F0-804067B83DFD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F5EB33D5-7C77-4D90-B90D-E1A8E014A28E}
-	EndGlobalSection
-EndGlobal
