--- UI/PingMonitor/PingMonitor/App.xaml
@@ -1,24 +1,0 @@
-﻿<Application x:Class="HuskyRobotics.UI.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:HuskyRobotics.UI"
-             xmlns:sys="clr-namespace:System;assembly=mscorlib"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-        <Thickness x:Key="DefaultMargin" Left="4" Right="4" Top="4" Bottom="4"/>
-        <Thickness x:Key="DefaultPadding" Left="4" Right="4" Top="4" Bottom="4"/>
-
-        <Style TargetType="Button">
-            <Setter Property="Margin" Value="{StaticResource ResourceKey=DefaultMargin}"></Setter>
-            <Setter Property="Padding" Value="{StaticResource ResourceKey=DefaultPadding}"></Setter>
-        </Style>
-        <Style TargetType="TextBox">
-            <Setter Property="Margin" Value="{StaticResource ResourceKey=DefaultMargin}"></Setter>
-            <Setter Property="Padding" Value="{StaticResource ResourceKey=DefaultPadding}"></Setter>
-        </Style>
-        <Style TargetType="ListView">
-            <Setter Property="Margin" Value="{StaticResource ResourceKey=DefaultMargin}"></Setter>
-            <Setter Property="Padding" Value="{StaticResource ResourceKey=DefaultPadding}"></Setter>
-        </Style>
-    </Application.Resources>
-</Application>
