--- UI/PingMonitor/PingMonitor/PingMonitor.xaml
@@ -1,24 +1,0 @@
-﻿<UserControl x:Class="HuskyRobotics.UI.PingMonitor"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:HuskyRobotics.UI"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Grid>
-        <DockPanel>
-            <StackPanel DockPanel.Dock="Bottom" Orientation="Horizontal" Margin="4">
-                <TextBlock Text="Connections: " />
-                <TextBlock Text="{Binding PingResults.Count}" />
-            </StackPanel>
-            <!-- DataGrid will generate columns based on the properties of it's rows by default -->
-            <DataGrid ItemsSource="{Binding PingResults}" AutoGenerateColumns="False">
-                <DataGrid.Columns>
-                    <DataGridTextColumn Binding="{Binding Address}" Header="Address" />
-                    <DataGridTextColumn Binding="{Binding ResponseTimeMs}" Header="Response Time (ms)" />
-                </DataGrid.Columns>
-            </DataGrid>
-        </DockPanel>
-    </Grid>
-</UserControl>
